requirements.txt
alembic==1.14.0
fastapi==0.115.7
pydantic==2.10.6
pytest==8.3.4
python_jose==3.3.0
SQLAlchemy==2.0.36
uvicorn==0.34.0
pymysql==1.1.1